python can handle index and evaluation tasks

+++ GPU_server/ir_collection_handler.c
@@ -84,7 +84,7 @@ int getCollection(Collection *collection){
   );
 
   printf("Loading postings\n");
-  int postingsSize = sizeof(PostingFreq) * collection->terms;
+  size_t postingsSize = sizeof(PostingFreq) * collection->terms;
   printf("Postings size: %.6f mb\n", (double) postingsSize / 1000000);
   PostingFreq *postingsFreq = (PostingFreq *) malloc(postingsSize);
   int status = getPostingsBin(
